walking: draft of velocity reference follower
Basics are there, seems to work ok, though
not tested in depth yet.
The following behavior is fully proportional
(no additional discretization)
and in three axes (x, y, and z rotation)
that can be freely combined
(e.g. walking forward while rotating is supported).
The foot profile is just rectangular for now,
and as result the motion doesn't look that elegant yet.