sim: doc + increase speed
The simulation now runs close to real time (> 85 %) on my laptop
(i7-6700HQ + Intel HD Graphics 530) when the Gazebo GUI
is kept small. Most of the GPU load is just rendering the GUI
(with the 3D viewport). Using my laptop's Nvidia GPU (GTX 960M)
gave limited improvements, likely because the CPU
became the bottleneck.
Included two lines in the docker-compose.yml, to be enabled
when using an Nvidia GPU.